Divides code into separate modules
